managesieve: Add a socket timeout parameter.
Unfortunately `settimeout()` puts the socket in non-blocking mode. Thus it is
then no longer possible to open the socket as a file and perform communication
through that file. The timeout is then done in `recv[]` (this is documented in
https://docs.python.org/3/library/socket.html#socket-timeouts). Therefore I
had to change the line-splitting logic for unencrypted connections.

Also, due to removed file-like access, `SSLFakeFile` is no longer needed.

Co-authored-by: Hartmut Goebel <h.goebel@crazy-compilers.com>