sieveshell: Keep line-endings on file IO.
When writing the script to this temporary file, Python's 'universal newlines
mode' is enabled by default. This will convert CRLF line endings to LF on
Linux. Anyhow CRLF should be preserved and might be useful when
e.g. specifying the message body of a vacation statement. (Actually RFC2822
requires CRLF in email messages, so some mail-servers might require it.)